LayoutTests/fast/dom/Element/id-in-formcollection.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<form id="idf1">
<select id="ids1" name="name1"><option selected>FAILURE</option></select>
<select id="ids1" name="name2"><option selected>FAILURE</option></select>
</form>

<script>
description("Test using id in HTMLFormCollection");

var elems = document.getElementById("idf1").elements;
shouldBe('elems["ids1"][1].getAttribute("name")', '"name2"');

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
