third_party/blink/web_tests/fast/dom/Element/id-in-formcollection-expected.txt

Test using id in HTMLFormCollection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elems["ids1"][1].getAttribute("name") is "name2"
PASS successfullyParsed is true

TEST COMPLETE

FAILURE
 
FAILURE
